pyproject.toml
[project]
name = "git-issue-to-markdown"
version = "0.1.0"
description = "CLI tool to fetch Gitea issues and add them to a markdown file"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "py-gitea",
]

[project.scripts]
git-issue-to-markdown = "git_issue_to_markdown.main:main"

[tool.uv.sources]
py-gitea = { path = "D:\\GIT\\GitHub\\py-gitea", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_issue_to_markdown"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
